theme [nfc]: Pull `zulipThemeData` out from app.dart to new theme.dart file
